remove redundant conversions

+++ base_layer/core/src/mempool/proto/stats_response.rs
@@ -40,9 +40,9 @@ impl TryFrom<ProtoStatsResponse> for StatsResponse {
 impl From<StatsResponse> for ProtoStatsResponse {
     fn from(stats: StatsResponse) -> Self {
         Self {
-            total_txs: stats.total_txs as u64,
-            unconfirmed_txs: stats.unconfirmed_txs as u64,
-            reorg_txs: stats.reorg_txs as u64,
+            total_txs: stats.total_txs,
+            unconfirmed_txs: stats.unconfirmed_txs,
+            reorg_txs: stats.reorg_txs,
             total_weight: stats.total_weight,
         }
     }
